Features: Create, confirm, get order and get orders by farmer id

+++ modules/common/src/main/java/shared/constants/Constants.java
@@ -3,5 +3,5 @@ package shared.constants;
 public final class Constants {
     private Constants(){}
 
-    public static final Double UNIT_COST = 30000.0;
+    public static final Float UNIT_COST = 30000F;
 }
